modified module loading code to use the new low-memory spec

+++ include/globals.h
@@ -6,7 +6,7 @@
 #include <stdio.h>
 #include <time.h>
 
-#define NUM_CHANNELS ((uint8_t)4)
+#define NUM_CHANNELS 4
 #define NUM_LEDS 2
 #define V_MAX ((uint16_t)1023)
 #define HYSTERESIS ((uint8_t)3)
